feat: use dart bytecode instead of native lib.
