--- packages/kraken_websocket/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'kraken_websocket'
