--- packages/kraken_websocket/bridge/.gitignore
@@ -1,4 +1,0 @@
-.kraken-test-tools
-cmake-build-*
-kraken
-build
