+++ packages/kraken_websocket/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -7,12 +7,10 @@ import Foundation
 
 import connectivity_macos
 import kraken
-import kraken_websocket
 import shared_preferences_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   KrakenPlugin.register(with: registry.registrar(forPlugin: "KrakenPlugin"))
-  KrakenWebsocketPlugin.register(with: registry.registrar(forPlugin: "KrakenWebsocketPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
 }
