--- packages/kraken_websocket/ios/Classes/KrakenWebsocketPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface KrakenWebsocketPlugin : NSObject<FlutterPlugin>
-@end
