--- packages/kraken_websocket/ios/Classes/KrakenWebsocketPlugin.m
@@ -1,20 +1,0 @@
-#import "KrakenWebsocketPlugin.h"
-
-@implementation KrakenWebsocketPlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  FlutterMethodChannel* channel = [FlutterMethodChannel
-      methodChannelWithName:@"kraken_websocket"
-            binaryMessenger:[registrar messenger]];
-  KrakenWebsocketPlugin* instance = [[KrakenWebsocketPlugin alloc] init];
-  [registrar addMethodCallDelegate:instance channel:channel];
-}
-
-- (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
-  if ([@"getPlatformVersion" isEqualToString:call.method]) {
-    result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
-  } else {
-    result(FlutterMethodNotImplemented);
-  }
-}
-
-@end
