+++ packages/kraken_websocket/lib/kraken_websocket.dart
@@ -1,21 +1,12 @@
 import 'websocket_module.dart';
 import 'package:kraken/module.dart';
-import 'platform.dart';
-import 'dart:ffi';
-
-typedef Native_InitBridge = Void Function();
-typedef Dart_InitBridge = void Function();
-
-final Dart_InitBridge _initBridge =
-nativeDynamicLibrary.lookup<NativeFunction<Native_InitBridge>>('initBridge').asFunction();
-
-void initBridge() {
-  _initBridge();
-}
+import 'package:kraken/bridge.dart';
+import 'websocket_qjsc.dart';
 
 class KrakenWebsocket {
   static void initialize() {
-    initBridge();
+    print(byteData);
+    registerPluginByteCode(byteData, 'kraken_websocket');
     ModuleManager.defineModule((moduleNamager) => WebSocketModule(moduleNamager));
   }
 }
