+++ packages/kraken_websocket/lib/websocket_qjsc.dart
@@ -1,0 +1,7 @@
+
+import 'dart:typed_data';
+
+Uint8List byteData = Uint8List.fromList([
+  1,65,22,105,110,116,101,114,110,97,108,58,47,47,22,118,97,108,105,100,97,116,101,85,114,108,70,105,110,105,116,80,114,111,112,101,114,116,121,72,97,110,100,108,101,114,115,70,111,114,69,118,101,110,116,84,97,114,103,101,116,115,20,82,101,97,100,121,83,116,97,116,101,20,66,105,110,97,114,121,84,121,112,101,22,119,115,67,108,105,101,110,116,77,97,112,44,100,105,115,112,97,116,99,104,87,101,98,83,111,99,107,101,116,69,118,101,110,116,30,98,117,105,108,116,73,110,69,118,101,110,116,115,36,49,18,87,101,98,83,111,99,107,101,116,8,111,112,101,110,10,99,108,111,115,101,10,101,114,114,111,114,22,69,118,101,110,116,84,97,114,103,101,116,32,97,100,100,69,118,101,110,116,76,105,115,116,101,110,101,114,8,115,101,110,100,12,107,114,97,107,101,110,46,97,100,100,75,114,97,107,101,110,77,111,100,117,108,101,76,105,115,116,101,110,101,114,6,117,114,108,16,112,114,111,116,111,99,111,108,18,115,117,98,115,116,114,105,110,103,14,105,110,100,101,120,79,102,2,58,4,119,115,6,119,115,115,162,1,70,97,105,108,101,100,32,116,111,32,99,111,110,115,116,114,117,99,116,32,39,87,101,98,83,111,99,107,101,116,39,58,32,84,104,101,32,85,82,76,39,115,32,115,99,104,101,109,101,32,109,117,115,116,32,98,101,32,101,105,116,104,101,114,32,39,119,115,39,32,111,114,32,39,119,115,115,39,46,32,39,34,39,32,105,115,32,110,111,116,32,97,108,108,111,119,101,100,46,22,101,118,101,110,116,84,97,114,103,101,116,26,98,117,105,108,116,73,110,69,118,101,110,116,115,14,95,108,111,111,112,95,49,2,105,18,101,118,101,110,116,78,97,109,101,24,112,114,111,112,101,114,116,121,78,97,109,101,4,111,110,2,95,38,114,101,109,111,118,101,69,118,101,110,116,76,105,115,116,101,110,101,114,20,67,79,78,78,69,67,84,73,78,71,8,79,80,69,78,14,67,76,79,83,73,78,71,12,67,76,79,83,69,68,8,98,108,111,98,22,97,114,114,97,121,98,117,102,102,101,114,16,99,108,105,101,110,116,73,100,10,101,118,101,110,116,12,99,108,105,101,110,116,20,114,101,97,100,121,83,116,97,116,101,32,99,111,110,110,101,99,116,105,111,110,83,116,97,116,117,115,8,116,121,112,101,12,99,108,111,115,101,100,26,101,115,116,97,98,108,105,115,104,109,101,110,116,14,99,111,110,115,111,108,101,50,87,101,98,83,111,99,107,101,116,32,99,111,110,110,101,99,116,105,111,110,32,116,111,32,39,20,39,32,102,97,105,108,101,100,58,32,40,69,114,114,111,114,32,105,110,32,99,111,110,110,101,99,116,105,111,110,32,4,58,32,26,100,105,115,112,97,116,99,104,69,118,101,110,116,20,101,120,116,101,110,115,105,111,110,115,20,98,105,110,97,114,121,84,121,112,101,24,105,110,118,111,107,101,77,111,100,117,108,101,8,105,110,105,116,4,105,100,16,99,97,108,108,98,97,99,107,16,97,100,100,69,118,101,110,116,8,99,111,100,101,20,109,111,100,117,108,101,78,97,109,101,8,100,97,116,97,14,0,6,0,160,1,0,1,0,1,0,1,5,1,162,1,0,0,0,190,0,236,203,40,164,3,1,4,0,0,138,2,14,67,6,0,0,0,10,0,6,0,10,190,1,10,166,3,0,0,64,168,3,0,1,64,170,3,0,2,64,172,3,0,3,64,174,3,1,0,112,176,3,0,4,64,178,3,1,5,112,180,3,2,10,48,236,1,3,8,112,180,3,1,7,32,190,0,199,190,1,200,190,4,193,5,97,9,0,97,6,0,97,4,0,190,2,197,17,233,4,14,11,205,237,14,190,3,198,17,233,4,14,11,206,237,14,11,193,4,4,219,0,0,0,4,220,0,0,0,4,51,0,0,0,4,221,0,0,0,38,4,0,193,6,97,7,0,56,222,0,0,0,97,8,0,189,5,86,218,0,0,0,1,190,6,84,223,0,0,0,0,190,7,84,224,0,0,0,0,190,8,84,220,0,0,0,0,6,193,8,14,194,7,104,8,0,193,9,56,225,0,0,0,66,226,0,0,0,190,9,36,1,0,14,56,143,0,0,0,66,101,0,0,0,56,138,0,0,0,4,218,0,0,0,11,10,76,63,0,0,0,9,76,62,0,0,0,9,76,61,0,0,0,98,9,0,76,64,0,0,0,36,3,0,41,164,3,1,34,0,19,76,0,11,12,59,0,3,76,28,93,14,0,19,44,0,8,8,0,8,8,43,59,0,10,8,34,108,33,33,33,43,18,14,67,6,0,166,3,1,1,1,6,0,0,82,2,198,3,0,1,0,200,3,1,0,32,97,0,0,207,66,229,0,0,0,179,207,66,230,0,0,0,4,231,0,0,0,36,1,0,36,2,0,199,98,0,0,4,232,0,0,0,172,232,43,98,0,0,4,233,0,0,0,172,232,32,56,145,0,0,0,17,4,234,0,0,0,66,92,0,0,0,98,0,0,4,235,0,0,0,36,2,0,33,1,0,47,41,164,3,2,5,18,128,113,153,8,14,67,6,0,168,3,2,2,2,2,0,1,25,4,216,3,0,1,64,218,3,0,1,64,220,3,0,0,0,222,3,0,1,0,190,0,77,238,0,0,0,199,179,200,196,208,231,163,232,9,195,196,237,14,147,1,234,243,41,164,3,9,7,0,0,36,43,43,23,23,14,67,6,0,0,1,2,1,6,2,2,53,3,222,3,0,1,0,224,3,0,0,64,226,3,0,1,64,218,3,1,3,216,3,0,3,219,207,71,199,4,242,0,0,0,195,157,200,56,143,0,0,0,66,101,0,0,0,220,196,11,190,0,77,65,0,0,0,76,65,0,0,0,190,1,77,66,0,0,0,76,66,0,0,0,36,3,0,41,164,3,10,9,3,23,43,70,0,12,18,63,18,14,67,6,0,0,0,1,0,3,1,0,12,1,16,0,1,0,226,3,1,1,8,199,195,4,243,0,0,0,219,157,71,40,164,3,14,1,13,14,67,6,0,0,1,1,1,6,2,0,52,2,128,1,0,1,0,16,0,1,0,224,3,0,1,226,3,1,1,8,199,207,7,169,232,23,195,66,244,0,0,0,219,195,4,243,0,0,0,220,157,71,36,2,0,14,234,13,195,66,223,0,0,0,219,207,36,2,0,14,195,4,243,0,0,0,220,157,207,73,41,164,3,17,6,13,28,103,14,64,53,14,67,6,0,0,1,0,1,5,0,0,69,1,170,3,0,1,0,207,207,4,245,0,0,0,113,179,22,73,4,245,0,0,0,73,207,207,4,246,0,0,0,113,180,22,73,4,246,0,0,0,73,207,207,4,247,0,0,0,113,181,22,73,4,247,0,0,0,73,207,207,4,248,0,0,0,113,182,22,73,4,248,0,0,0,73,41,164,3,34,5,3,88,88,88,88,14,67,6,0,0,1,0,1,3,0,0,25,1,172,3,0,1,0,207,4,249,0,0,0,4,249,0,0,0,73,207,4,250,0,0,0,4,250,0,0,0,73,41,164,3,42,3,3,63,63,14,67,6,0,176,3,2,3,2,5,2,0,132,2,5,246,3,0,1,0,248,3,0,1,0,250,3,1,0,32,252,3,3,1,32,254,3,4,2,32,174,3,4,13,170,3,2,1,97,0,0,101,0,0,207,71,199,98,0,0,105,246,0,0,0,97,1,0,98,0,0,65,254,0,0,0,200,208,65,0,1,0,0,97,2,0,17,4,219,0,0,0,171,232,15,220,65,246,0,0,0,17,99,1,0,14,235,175,0,17,4,220,0,0,0,171,232,15,220,65,248,0,0,0,17,99,1,0,14,235,152,0,17,4,221,0,0,0,171,105,142,0,0,0,220,65,248,0,0,0,17,99,1,0,14,191,201,98,1,0,17,220,65,248,0,0,0,171,232,13,4,1,1,0,0,17,99,2,0,14,234,43,17,220,65,246,0,0,0,171,232,13,4,2,1,0,0,17,99,2,0,14,234,21,17,220,65,245,0,0,0,171,232,11,4,2,1,0,0,17,99,2,0,14,14,56,3,1,0,0,66,221,0,0,0,4,4,1,0,0,98,0,0,65,227,0,0,0,157,4,5,1,0,0,157,4,6,1,0,0,157,98,2,0,157,4,7,1,0,0,157,208,65,221,0,0,0,157,36,1,0,14,14,98,0,0,98,1,0,67,254,0,0,0,98,0,0,66,8,1,0,0,208,36,1,0,14,41,164,3,48,29,18,33,58,48,48,48,58,18,48,58,18,63,58,13,18,53,53,14,53,53,14,53,55,8,178,114,8,58,69,14,198,7,1,0,2,3,2,6,7,0,213,1,5,198,3,0,1,0,200,3,0,1,0,228,1,0,1,0,226,1,0,1,0,16,0,1,32,236,1,8,13,170,3,2,1,172,3,3,1,166,3,0,1,174,3,4,13,168,3,1,1,178,3,6,13,12,2,199,12,3,200,97,2,0,43,195,52,196,33,0,0,17,100,2,0,101,0,0,17,232,8,98,2,0,27,36,0,0,14,14,98,2,0,220,65,245,0,0,0,67,245,0,0,0,98,2,0,220,65,246,0,0,0,67,246,0,0,0,98,2,0,220,65,247,0,0,0,67,247,0,0,0,98,2,0,220,65,248,0,0,0,67,248,0,0,0,98,2,0,191,67,9,1,0,0,98,2,0,191,67,228,0,0,0,98,2,0,221,65,249,0,0,0,67,10,1,0,0,222,207,237,14,98,2,0,207,67,227,0,0,0,98,2,0,220,65,245,0,0,0,67,254,0,0,0,98,2,0,56,225,0,0,0,66,11,1,0,0,4,218,0,0,0,4,12,1,0,0,207,36,3,0,67,13,1,0,0,101,4,0,98,2,0,65,13,1,0,0,113,98,2,0,73,94,5,0,98,2,0,101,6,0,238,14,98,2,0,40,164,3,90,15,54,128,73,73,73,73,48,48,74,23,48,73,163,83,58,14,74,7,1,0,2,2,2,6,0,0,54,4,128,4,0,1,0,156,4,0,1,0,16,0,1,0,230,1,0,1,0,12,4,200,8,199,56,225,0,0,0,66,11,1,0,0,4,218,0,0,0,4,15,1,0,0,195,65,13,1,0,0,207,38,2,0,36,3,0,14,195,196,52,4,223,0,0,0,71,207,208,36,2,0,41,164,3,108,3,28,173,73,14,66,7,1,0,1,1,1,6,0,0,36,2,102,0,1,0,16,0,1,0,8,199,56,225,0,0,0,66,11,1,0,0,4,218,0,0,0,4,224,0,0,0,195,65,13,1,0,0,207,38,2,0,36,3,0,41,164,3,113,2,13,168,14,66,7,1,0,2,1,2,7,1,0,49,3,160,4,0,1,0,146,2,0,1,0,16,0,1,0,170,3,2,1,8,199,195,219,65,247,0,0,0,67,254,0,0,0,56,225,0,0,0,66,11,1,0,0,4,218,0,0,0,4,220,0,0,0,195,65,13,1,0,0,207,208,38,3,0,36,3,0,41,164,3,116,3,13,63,173,14,67,6,0,0,3,0,3,3,1,0,15,3,162,4,0,1,0,248,3,0,1,0,164,4,0,1,0,176,3,5,1,207,4,218,0,0,0,171,232,6,219,209,208,238,14,41,164,3,122,3,3,48,29
+]);
+
