--- packages/kraken_websocket/macos/Classes/KrakenWebsocketPlugin.swift
@@ -1,19 +1,0 @@
-import Cocoa
-import FlutterMacOS
-
-public class KrakenWebsocketPlugin: NSObject, FlutterPlugin {
-  public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "kraken_websocket", binaryMessenger: registrar.messenger)
-    let instance = KrakenWebsocketPlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
-  }
-
-  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    switch call.method {
-    case "getPlatformVersion":
-      result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
-    default:
-      result(FlutterMethodNotImplemented)
-    }
-  }
-}
