--- packages/kraken_websocket/macos/libkraken_websocket.dylib
@@ -1,1 +1,0 @@
-../bridge/build/macos/libkraken_websocket.dylib
